requirements.txt
dash==2.16.1
plotly==5.20.0
pandas==2.2.1
streamlit==1.32.2
extra-streamlit-components==0.1.71
streamlit-authenticator==0.3.2
streamlit-option-menu
SQLAlchemy==2.0.31
psycopg2-binary
PyYAML
openpyxl
